Update lwgps.h
Reverted back to uint8_t satellite prn.

+++ lwgps/src/include/lwgps/lwgps.h
@@ -63,7 +63,7 @@ typedef float lwgps_float_t;
  * \brief           Satellite descriptor
  */
 typedef struct {
-    uint16_t num;      /*!< Satellite number */
+    uint8_t num;       /*!< Satellite number */
     uint8_t elevation; /*!< Elevation in degrees */
     uint16_t azimuth;  /*!< Azimuth in degrees */
     uint8_t snr;       /*!< Signal-to-noise ratio */
